Initial commit - Rewrite EZAvatarSetup (WIP)
First commit for the rewriting of EZAvatarSetup to improve it's functionality and provide additional usability. This is the first iteration of a new UI that allows more customization, creating categories and slotting in the materials for such category (i.e: sweater, hat, etc). As well as the new feature of creating animation clips for toggles.